requirements.txt
tensorflow = 2.3.0
pyqt5
pillow
opencv-python